+++ src/JK.Common.FluentValidation/Validators/NullValidator.cs
@@ -2,6 +2,10 @@
 
 namespace JK.Common.FluentValidation.Validators;
 
+/// <summary>
+/// Null object pattern AbstractValidator implementation.
+/// </summary>
+/// <typeparam name="T"></typeparam>
 public class NullValidator<T> : AbstractValidator<T>
 {
 }