+++ src/JK.Common/Extensions/TypeExtensions.cs
@@ -4,6 +4,9 @@ using JK.Common.TypeHelpers;
 
 namespace JK.Common.Extensions;
 
+/// <summary>
+/// Helper and utility extension methods for <see cref="Type"/>.
+/// </summary>
 public static class TypeExtensions
 {
     /// <summary>
@@ -20,6 +23,11 @@ public static class TypeExtensions
             : throw new ArgumentException("Only interfaces can be passed as T.");
     }
 
+    /// <summary>
+    /// Attempts to get the underlying <see cref="Type"/> from an Entity Framework proxied type.
+    /// </summary>
+    /// <param name="type">Suspected DynamicProxy type.</param>
+    /// <returns>Underlying entity <see cref="Type"/>.</returns>
     public static Type GetTypeFromEntity(this Type type)
     {
         if (type.BaseType is not null && type.Namespace == "System.Data.Entity.DynamicProxies")
@@ -30,9 +38,16 @@ public static class TypeExtensions
     }
 
     /// <summary>
+    /// Determines whether or not a type is nullable (including <see cref="Nullable{T}"/>, aka T?)
+    /// </summary>
+    /// <param name="type">The type to check</param>
+    /// <returns>True if nullable, otherwise false</returns>
+    public static bool IsNullable(this Type type) => TypeHelper.IsNullable(type);
+
+    /// <summary>
     /// Determines whether or not a type uses <see cref="Nullable{T}"/>, aka T?
     /// </summary>
     /// <param name="type">The type to check</param>
     /// <returns>True if <see cref="Nullable{T}"/>, otherwise false</returns>
-    public static bool IsNullable(this Type type) => TypeHelper.IsNullable(type);
+    public static bool IsNullableT(this Type type) => TypeHelper.IsNullableT(type);
 }