+++ src/JK.Common/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.com/jeremyknight-me).
 
+## Unreleased
+
+## Added
+
+- `IsNullable` helper method in `TypeHelper`
+
+## Changed
+
+- `IsNullable`  to `IsNullableT` int `TypeHelper`.
+
 ## 5.2.0
 
 ### Added